+++ Jellyfin.Plugin.TuneIn.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26730.3
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Jellyfin.Plugin.TuneIn", "Jellyfin.Plugin.TuneIn\Jellyfin.Plugin.TuneIn.csproj", "{EF7A1494-46BD-4125-A66D-724A2D225FA0}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{EF7A1494-46BD-4125-A66D-724A2D225FA0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EF7A1494-46BD-4125-A66D-724A2D225FA0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EF7A1494-46BD-4125-A66D-724A2D225FA0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EF7A1494-46BD-4125-A66D-724A2D225FA0}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {2F43363A-6AEB-4C25-B1E6-A0721EE68F97}
+	EndGlobalSection
+EndGlobal
