+++ Jellyfin.Plugin.TuneIn/Plugin.cs
@@ -1,0 +1,62 @@
+﻿using System;
+using System.Collections.Generic;
+using MediaBrowser.Common.Configuration;
+using MediaBrowser.Common.Plugins;
+using MediaBrowser.Model.Plugins;
+using MediaBrowser.Model.Serialization;
+using Jellyfin.Plugin.TuneIn.Configuration;
+
+namespace Jellyfin.Plugin.TuneIn
+{
+    /// <summary>
+    /// Class Plugin
+    /// </summary>
+    public class Plugin : BasePlugin<PluginConfiguration>, IHasWebPages
+    {
+        
+        public Plugin(IApplicationPaths applicationPaths, IXmlSerializer xmlSerializer)
+            : base(applicationPaths, xmlSerializer)
+        {
+            Instance = this;
+        }
+
+        public IEnumerable<PluginPageInfo> GetPages()
+        {
+            return new[]
+            {
+                new PluginPageInfo
+                {
+                    Name = "tunein",
+                    EmbeddedResourcePath = GetType().Namespace + ".Configuration.configPage.html"
+                }
+            };
+        }
+
+        private Guid _id = new Guid("10b82431-f939-4507-a8e2-9509dd3fd6ae");
+        public override Guid Id
+        {
+            get { return _id; }
+        }
+
+        /// <summary>
+        /// Gets the name of the plugin
+        /// </summary>
+        /// <value>The name.</value>
+        public override string Name
+        {
+            get { return "TuneIn"; }
+        }
+
+        /// <summary>
+        /// Gets the description.
+        /// </summary>
+        /// <value>The description.</value>
+        public override string Description => "TuneIn streams in your collection.";
+
+        /// <summary>
+        /// Gets the instance.
+        /// </summary>
+        /// <value>The instance.</value>
+        public static Plugin Instance { get; private set; }
+    }
+}
