--- Jellyfin.Plugins.TuneIn/Configuration/configPage.html
@@ -1,78 +1,0 @@
-﻿<!DOCTYPE html>
-<html>
-<head>
-    <title>TuneIn</title>
-</head>
-<body>
-    <div data-role="page" class="page type-interior pluginConfigurationPage TuneInConfigurationPage" data-require="emby-input,emby-button">
-
-        <div data-role="content">
-            <div class="content-primary">
-
-                <h1>Tune In</h1>
-
-                <form class="TuneInConfigurationForm">
-
-                    <div class="inputContainer">
-                        <input type="text" is="emby-input" id="txtUsername" label="Your Tune In username:"/>
-                    </div>
-                    <div class="inputContainer">
-                        <input type="text" is="emby-input" id="txtLatLon" label="Your LatLon:"/>
-                        <div class="fieldDescription">
-                            Your LatLon can be found at <a target="_blank" href="http://dbsgeo.com/latlon">http://dbsgeo.com/latlon</a>.
-                        </div>
-                    </div>
-
-                    <br/>
-                    <div>
-                        <button is="emby-button" type="submit" class="raised button-submit block"><span>Save</span></button>
-                    </div>
-                </form>
-            </div>
-        </div>
-
-        <script type="text/javascript">
-
-            var TuneInConfigurationPage = {
-                pluginUniqueId: "10b82431-f939-4507-a8e2-9509dd3fd6ae"
-            };
-
-            $('.TuneInConfigurationPage').on('pageshow', function (event) {
-
-                Dashboard.showLoadingMsg();
-
-                var page = this;
-
-                ApiClient.getPluginConfiguration(TuneInConfigurationPage.pluginUniqueId).then(function (config) {
-
-                    $('#txtUsername', page).val(config.Username || "");
-                    $('#txtLatLon', page).val(config.LatLon || "");
-
-                    Dashboard.hideLoadingMsg();
-                });
-            });
-
-            $('.TuneInConfigurationForm').on('submit', function (e) {
-
-                Dashboard.showLoadingMsg();
-
-                var form = this;
-
-                ApiClient.getPluginConfiguration(TuneInConfigurationPage.pluginUniqueId).then(function (config) {
-
-                    var Username = $('#txtUsername', form).val();
-                    var LatLon = $('#txtLatLon', form).val();
-
-                    config.Username = Username ? Username : null;
-                    config.LatLon = LatLon ? LatLon : null;
-
-                    ApiClient.updatePluginConfiguration(TuneInConfigurationPage.pluginUniqueId, config).then(Dashboard.processPluginConfigurationUpdateResult);
-                });
-
-                // Disable default form submission
-                return false;
-            });
-        </script>
-    </div>
-</body>
-</html>
