--- Jellyfin.Plugins.TuneIn/Jellyfin.Plugins.TuneIn.csproj
@@ -1,45 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <AssemblyVersion>1.2.0.0</AssemblyVersion>
-    <FileVersion>1.2.0.0</FileVersion>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <None Remove="Configuration\configPage.html" />
-    <None Remove="Images\backdrop.png" />
-    <None Remove="Images\primary.png" />
-    <None Remove="Images\thumb.png" />
-    <None Remove="Images\tunein-bylanguage.png" />
-    <None Remove="Images\tunein-bylocation.png" />
-    <None Remove="Images\tunein-localradio.png" />
-    <None Remove="Images\tunein-music.png" />
-    <None Remove="Images\tunein-myfavs.png" />
-    <None Remove="Images\tunein-podcasts.png" />
-    <None Remove="Images\tunein-sports.png" />
-    <None Remove="Images\tunein-talk.png" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <EmbeddedResource Include="Configuration\configPage.html" />
-    <EmbeddedResource Include="Images\backdrop.png" />
-    <EmbeddedResource Include="Images\primary.png" />
-    <EmbeddedResource Include="Images\thumb.png" />
-    <EmbeddedResource Include="Images\tunein-bylanguage.png" />
-    <EmbeddedResource Include="Images\tunein-bylocation.png" />
-    <EmbeddedResource Include="Images\tunein-localradio.png" />
-    <EmbeddedResource Include="Images\tunein-music.png" />
-    <EmbeddedResource Include="Images\tunein-myfavs.png" />
-    <EmbeddedResource Include="Images\tunein-podcasts.png" />
-    <EmbeddedResource Include="Images\tunein-sports.png" />
-    <EmbeddedResource Include="Images\tunein-talk.png" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="HtmlAgilityPack" Version="1.8.14" />
-    <PackageReference Include="Jellyfin.Controller" Version="10.1.0" />
-    <PackageReference Include="System.Memory" Version="4.5.1" />
-  </ItemGroup>
-
-</Project>
